feat(frontend): add Redux store with a sessions reducer
We'll need this to maintain API resources requested during a
user or viewers session on the website.

Start off with sessionReducer, which is intended to maintain
authenticated session state. By default, it does nothing,
but it allows us to wrap our application in a react-redux
`Provider`, so Redux will be wired up for the entire application.

Signed-off-by: Kevin Morris <kevr@0cost.org>